feat: 상품 state 위치 변경
